move format_bytes to utils
